changed scope from deceased to player
Early on the scope player gets defined, which by my understanding is the same as deceased. But other then deceased is not only available in the gold path but also in the title path. Therefore eliminating the error that happens in line 5 of this file when you pick underdog as heir choice.